Move can-derived bodies to their own module
